pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "churn-ml"
version = "0.1.0"
description = "Telco customer churn — MLOps portfolio pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "lightgbm>=4.0",
    "matplotlib>=3.7",
    "mlflow>=2.10",
    "pandas>=2.0",
    "pydantic>=2.6",
    "pandera>=0.20",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
]

[project.scripts]
churn-validate = "churn_ml.data.validate:cli"
churn-eda = "churn_ml.eda.run:cli"
churn-split = "churn_ml.data.split:cli"
churn-features = "churn_ml.features.run:cli"
churn-train-baseline = "churn_ml.models.run_baseline:cli"
churn-train-lightgbm = "churn_ml.models.run_lightgbm:cli"
churn-evaluate = "churn_ml.evaluation.run:cli"
churn-batch-predict = "churn_ml.batch_predict.run:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
